+++ source/dotnet/Services/Infrastructure/Integration/ICalculationResultReadyIntegrationEventFactory.cs
@@ -12,21 +12,13 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-namespace Energinet.DataHub.Wholesale.Application;
-
 using Energinet.DataHub.Wholesale.Application.Processes.Model;
+using Energinet.DataHub.Wholesale.Contracts.Events;
 using Energinet.DataHub.Wholesale.Domain.ProcessStepResultAggregate;
 
-namespace Energinet.DataHub.Wholesale.Application;
+namespace Energinet.DataHub.Wholesale.Infrastructure.Integration;
 
-public interface IIntegrationEventPublisher
+public interface ICalculationResultReadyIntegrationEventFactory
 {
-    Task PublishAsync(ProcessCompletedEventDto processCompletedEvent);
-
-    Task PublishAsync(ProcessStepResult processStepResultDto, ProcessCompletedEventDto processCompletedEventDto);
-
-    /// <summary>
-    /// Fetches outbox messages and publishes them as integration events.
-    /// </summary>
-    Task PublishIntegrationEventsAsync(CancellationToken token);
+    CalculationResultCompleted CreateCalculationResultCompletedForGridArea(ProcessStepResult processStepResultDto, ProcessCompletedEventDto processCompletedEventDto);
 }